misc-coding/ai-research-platform/sub-projects/twin-report-kb/docker/diff_worker/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
asyncpg==0.29.0
pydantic==2.5.0
httpx==0.25.2
python-multipart==0.0.6

# Text processing and diff analysis
# difflib is built into Python, no separate package needed
python-Levenshtein==0.23.0
fuzzywuzzy==0.18.0
nltk==3.8.1
spacy==3.7.2

# AI/ML libraries
transformers==4.36.0
sentence-transformers==2.2.2
torch==2.1.1

# Database and storage
sqlalchemy==2.0.23
alembic==1.13.0

# Background tasks
celery==5.3.4
redis==5.0.1

# Utilities
python-dotenv==1.0.0
structlog==23.2.0
tenacity==8.2.3
pyyaml==6.0.1 